Finish 4.21*: bloglist expansion, step9

+++ part4/controllers/blogs.js
@@ -35,6 +35,16 @@ blogRouter.post("/", async (request, response) => {
 });
 
 blogRouter.delete("/:id", async (request, response) => {
+    const decodedToken = jwt.verify(request.token, process.env.SECRET);
+    if (!decodedToken.id) {
+        return response.status(401).json({ error: "token invalid" });
+    }
+
+    const blog = await Blog.findById(request.params.id);
+
+    if (blog.user.toString() !== decodedToken.id) {
+        return response.status(401).json({ error: "invalid user" });
+    }
     await Blog.findByIdAndRemove(request.params.id);
     response.status(204).end();
 });
